+++ reapy/core/__init__.py
@@ -4,7 +4,7 @@ from .audio_accessor import AudioAccessor
 from .envelope import Envelope, EnvelopeList
 from .fx import FX, FXList, FXParam, FXParamsList
 from .item import CC, CCList, Item, Note, NoteList, Source, Take
-from .project import Marker, Project, Region, TimeSelection, get_project_by_name
+from .project import Marker, Project, Region, TimeSelection
 from .track import AutomationItem, Send, Track, TrackList
 from .window import MIDIEditor, ToolTip, Window
 
@@ -36,7 +36,6 @@ __all__ = [
     "Project",
     "Region",
     "TimeSelection",
-    "get_project_by_name",
     # core.track
     "AutomationItem",
     "Send",