+++ reapy/core/project/__init__.py
@@ -1,4 +1,4 @@
 from .marker import Marker
-from .project import Project, get_project_by_name
+from .project import Project
 from .region import Region
 from .time_selection import TimeSelection